requirements.txt
aiohttp
requests
python-dotenv
anthropic
playwright
playwright-stealth
